% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_construction.R
\name{add.variable}
\alias{add.variable}
\title{Add a variable to a call}
\usage{
add.variable(query, variable, name = deparse(substitute(variable)))
}
\arguments{
\item{query}{A URL containing authentication for EPA API}

\item{variable}{A variable for a call. Consult VARIABLE.TYPES for possible variables.}

\item{name}{Default argument should be left as is. Will take the name used for variable above
to create the final URL.}
}
\value{
A URL containing query + variable.
}
\description{
Add a variable to a call
}
\examples{
\dontrun{
endpoint <- 'dailyData/byState'
state <- "37"
call <- epair:::create.base.call(endpoint)
call <- add.variable(call, state)
call     # Call requires more variables before being placed
}
}
