% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_construction.R
\name{add.variables}
\alias{add.variables}
\title{Add variables to a query}
\usage{
add.variables(query, variables, name = NA)
}
\arguments{
\item{query}{A URL containing authentication for the EPA API site.}

\item{variables}{A list of variables. Each variable should be declared with the appropriate name.
Consult VARIABLE.TYPES for the right names.}

\item{name}{A list containing names of API variables.}
}
\value{
A URL consisting of query + variables.
}
\description{
Add variables to a query
}
\examples{
\dontrun{
endpoint <- "dailyData/byState"
variable.list <- list("state" = '37', 
                      "bdate" = '20200101', 
                      "edate" = '20200102', 
                      "param" = '44201')
call <- epair:::create.base.call(endpoint)
call <- add.variables(call, variable.list)
call
}
}
