% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/making_services.R
\name{assign.description.to.services}
\alias{assign.description.to.services}
\title{Assign a description to each service}
\usage{
assign.description.to.services(services)
}
\arguments{
\item{services}{A list of services offered by the EPA API.}
}
\value{
The list of services with descriptions for each service.
}
\description{
Assign a description to each service
}
\examples{
\dontrun{
services <- assign.description.to.services(services)
services[[1]]$Description
}
}
