% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/making_services.R
\name{change.classes.filter}
\alias{change.classes.filter}
\title{Update name to parameter class entry}
\usage{
change.classes.filter(services)
}
\arguments{
\item{services}{List of services offered by the API.}
}
\value{
Services with corrected name of filter for parameter classes.
}
\description{
Update name to parameter class entry
}
\examples{
services <- get.services()
services <- change.classes.filter( services )
services$List$Filter$`Parameter Classes (groups of parameters, like criteria or all)`
}
