% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_placement.R
\name{clear.all.cached}
\alias{clear.all.cached}
\title{Removes all cached memory of perform.call}
\usage{
clear.all.cached(directory = "/cache")
}
\arguments{
\item{directory}{Place inside user-level cache directory that was used
to store the cached data previously. Default: "/cache".}
}
\value{
'Done' if data was successfully forgotten, error message if 
cache directory was not found
}
\description{
Removes all cached memory of perform.call
}
\examples{
\dontrun{
clear.all.cached() 
}
}
