% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_construction.R
\name{create.base.call}
\alias{create.base.call}
\title{Make the first call when forming a query.}
\usage{
create.base.call(endpoint)
}
\arguments{
\item{endpoint}{Endpoint for forming a query. See ENDPOINTS for all available endpoints. See 
SERVICES if you know the service but not the endpoint.}
}
\value{
A URL string containing authentication for the call.
}
\description{
Make the first call when forming a query.
}
\examples{
\dontrun{
endpoint <- "list/states"
call <- epair:::create.base.call(endpoint)
call
}
}
