% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets.R
\docType{data}
\name{endpoints}
\alias{endpoints}
\title{Endpoints available in the EPA API}
\description{
The endpoints vector contains all endpoints available in the EPA API. 
To get endpoints directly from the site, use get.endpoints().
}
\keyword{datasets}
