% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epa_scrapers.R
\name{find.endpoints.in.tables}
\alias{find.endpoints.in.tables}
\title{Take a list of html tables from api and output all endpoints}
\usage{
find.endpoints.in.tables(list.tables)
}
\arguments{
\item{list.tables}{List of HTML tables from EPA API}
}
\value{
Vector with only endpoints for the API.
}
\description{
Take a list of html tables from api and output all endpoints
}
\examples{
\dontrun{
API.tables <- epair:::get.all.tables()
endpoints <- epair:::find.endpoints.in.tables(API.tables)
endpoints
}
}
