% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/making_services.R
\name{generate.filter.content}
\alias{generate.filter.content}
\title{Generate filter content for an API filter}
\usage{
generate.filter.content(i, df)
}
\arguments{
\item{i}{Row number at which to get the information for filter.}

\item{df}{The data frame containing filter information.}
}
\value{
A list with filter content (endpoint, required variables, etc.)
}
\description{
Generate filter content for an API filter
}
\examples{
\dontrun{
tbls <- get.all.tables()
single <- tbls[[7]]
content <- generate.filter.content(1, single)
}
}
