% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/making_services.R
\name{generate.filters.list}
\alias{generate.filters.list}
\title{Create a list of filters}
\usage{
generate.filters.list(df)
}
\arguments{
\item{df}{A data frame having filter information (e.g. name, required variables).}
}
\value{
A list containing filters and respective info.
}
\description{
Create a list of filters
}
\examples{
\dontrun{
tbls <- get.all.tables()
single <- tbls[[8]]
generate.filters.list(single)
}
}
