% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epa_scrapers.R
\name{get.all.tables}
\alias{get.all.tables}
\title{Get all the html tables in the API site}
\usage{
get.all.tables()
}
\value{
A list of HTML tables from the EPA API site.
}
\description{
Get all the html tables in the API site
}
\examples{
\dontrun{
html.tables.list <- get.all.table()
html.tables.list
}
}
