% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get.endpoints}
\alias{get.endpoints}
\title{Get all endpoints from EPA API}
\usage{
get.endpoints()
}
\value{
Vector of endpoints from the API
}
\description{
Get all endpoints from EPA API
}
\examples{
endpoints <- get.endpoints()
endpoints
}
