% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/making_services.R
\name{get.first.entries}
\alias{get.first.entries}
\title{Get first entries for filter names}
\usage{
get.first.entries(df)
}
\arguments{
\item{df}{Data frame with filters}
}
\value{
A vector of indices. These indices are where the first entry for a filter 
exists in the data frame.
}
\description{
Get first entries for filter names
}
\examples{
\dontrun{
tbls <- get.all.tables()
single <- tbls[[10]]
first.occurrences <- get.first.entries(single)
}
}
