% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/making_services.R
\name{get.first.entry.for.filter}
\alias{get.first.entry.for.filter}
\title{Get the first entry for a filter name}
\usage{
get.first.entry.for.filter(filter.name, df)
}
\arguments{
\item{filter.name}{Name of the filter in API}

\item{df}{Data frame containing filter info.}
}
\value{
The index for the first occurrence of the filter in the data frame.
}
\description{
Get the first entry for a filter name
}
\examples{
\dontrun{
tbls <- get.all.tables()
single <- tbls[[11]]
get.first.entry.for.filter("Filter Name", single)
}
}
