% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/making_services.R
\name{get.first.occurences}
\alias{get.first.occurences}
\title{Get first entries for filter names}
\usage{
get.first.occurences(df)
}
\arguments{
\item{df}{Data frame with filters}
}
\value{
A vector of indices. These indices are where the first entry for a filter 
exists in df.
}
\description{
Get first entries for filter names
}
\examples{
\dontrun{
tbls <- get.all.tables()
single <- tbls[[10]]
first.occurences <- get.first.occurences( single )
}
}
