% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epa_scrapers.R
\name{get.services}
\alias{get.services}
\title{Get a list of services the EPA API offers}
\usage{
get.services()
}
\value{
List of services the EPA API offers.
}
\description{
Get a list of services the EPA API offers
}
\examples{
\dontrun{
services <- epair:::get.services()
services
}
}
