% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epa_scrapers.R
\name{get.table}
\alias{get.table}
\title{Get an HTML table at URL}
\usage{
get.table(url, table.xpath)
}
\arguments{
\item{url}{URL to get table from}

\item{table.xpath}{The X path to the table}
}
\value{
A data frame of the HTML table
}
\description{
Get an HTML table at URL
}
\examples{
\dontrun{
url <- "https://aqs.epa.gov/aqsweb/documents/data_api.html"
table.path <- '//*[@id="main-content"]/div[2]/div[1]/div/div/table[1]'
df <- epair:::get.table(url, table.path)
df
}
}
