% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transforms.R
\name{get.transpose}
\alias{get.transpose}
\title{Transpose a data frame}
\usage{
get.transpose(df)
}
\arguments{
\item{df}{Data frame to be transposed}
}
\value{
The transposed data frame. First variable entries become 
column names.
}
\description{
Transpose a data frame
}
\examples{
service <- c("Sign up")
description <- c("Email will\r\n\t\t\t\t\t\t\t be sent to the registered 
address from aqsdatamart@epa.gov.")
df <- data.frame(service, description)
t.df <- epair:::get.transpose(df)
t.df
}
