% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/making_services.R
\name{get.unique.filters}
\alias{get.unique.filters}
\title{Get filter names in data frame}
\usage{
get.unique.filters(df)
}
\arguments{
\item{df}{Data frame containing repeated or mixed filter names.}
}
\value{
Vector containing only filter names for the API service. Service name 
and repeated filter names are removed.
}
\description{
Get filter names in data frame
}
\examples{
\dontrun{
tbls <- get.all.tables()
single <- tbls[[6]]
get.unique.filters(single)
}
}
