% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epa_scrapers.R
\name{get.variables}
\alias{get.variables}
\title{Populate variables for info on making requests}
\usage{
get.variables()
}
\value{
Data frame containing variables and information about them used in the EPA API.
}
\description{
Populate variables for info on making requests
}
\examples{
\dontrun{
vars <- epair:::get.variables()
vars$edate
}
}
