% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_service.R
\name{get_all_mas}
\alias{get_all_mas}
\title{Get Monitoring Agencies.}
\usage{
get_all_mas()
}
\value{
API response containing all Monitoring Agencies.
}
\description{
Get Monitoring Agencies.
}
\examples{
\dontrun{
mas <- get_all_mas()
mas$Data
}
}
