% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_service.R
\name{get_all_pqaos}
\alias{get_all_pqaos}
\title{Get Primary Quality Assurance Organizations.}
\usage{
get_all_pqaos()
}
\value{
API response containing all Primary Quality Assurance Organizations.
}
\description{
Get Primary Quality Assurance Organizations.
}
\examples{
\dontrun{
pqaos <- get_all_pqaos()
pqaos$Data
}
}
