% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annual_summary_service.R
\name{get_annual_summary_in_site}
\alias{get_annual_summary_in_site}
\title{Get annual summary data at a measurement site.}
\usage{
get_annual_summary_in_site(
  bdate,
  edate,
  param,
  state.fips,
  county,
  site,
  cached = TRUE,
  cache_directory = "/cache",
  cbdate = NULL,
  cedate = NULL
)
}
\arguments{
\item{bdate}{Beginning date to check.
Year, month, day format.}

\item{edate}{Ending date to check.
Year, month, day format.}

\item{param}{Pollutant parameter that site is measuring.}

\item{state.fips}{State FIPS code.
Use get_state_fips() if unsure.}

\item{county}{County code. 
Use get_counties_in_state() if unsure.}

\item{site}{Measurement site code.
Use get_sites_by_county() if unsure.}

\item{cached}{TRUE or FALSE specifying if the data from the call is to 
be cached. Default: TRUE. (Optional)}

\item{cache_directory}{Place inside user-level cache directory to store 
the cached data. Default: "/cache". (Optional)}

\item{cbdate}{Beginning date of last change to DB. (Optional)}

\item{cedate}{Ending date of last change to DB. (Optional)}
}
\value{
API response containing annual summary data for a site.
}
\description{
Get annual summary data at a measurement site.
}
\examples{
\dontrun{
bdate <- "20170618"
edate <- "20170618"
state.fips <- "37"
county <- "183"
site <- "0014"
param <- "44201"
result <- get_annual_summary_in_site(bdate,
                                     edate,
                                     param,
                                     state.fips,
                                     county,
                                     site)
result$Data
}
}
