% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_placement.R
\name{get_aqs_key}
\alias{get_aqs_key}
\title{Get an API key from the AQS API.}
\usage{
get_aqs_key(user.email)
}
\arguments{
\item{user.email}{Email provided by the user to get an API key for.}
}
\description{
Get an API key from the AQS API.
}
\examples{
\dontrun{
email <- "an.example.email@domain.com"
get_aqs_key(email)
}
}
