% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_service.R
\name{get_cbsas}
\alias{get_cbsas}
\title{Get all Core Based Statistical Areas.}
\usage{
get_cbsas()
}
\value{
API response containing a list of all Core Based Statistical Areas.
}
\description{
Get all Core Based Statistical Areas.
}
\examples{
\dontrun{
cbsas <- get_cbsas()
cbsas$Data
}
}
