% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_service.R
\name{get_counties_in_state}
\alias{get_counties_in_state}
\title{Get all counties within a state.}
\usage{
get_counties_in_state(state.fips)
}
\arguments{
\item{state.fips}{A state FIPS code. 
Use get_state_fips() to find the appropriate FIPS code.}
}
\value{
API response containing all counties and 
county codes within a given state.
}
\description{
Get all counties within a state.
}
\examples{
\dontrun{
state <- "37"
counties.in.state <- get_state_fips(state)
counties.in.state$Data
}
}
