% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_summary_service.R
\name{get_daily_summary_in_bbox}
\alias{get_daily_summary_in_bbox}
\title{Returns daily summary data given a bounding box (lat, long).}
\usage{
get_daily_summary_in_bbox(
  bdate,
  edate,
  param,
  minlat,
  maxlat,
  minlong,
  maxlong,
  cbdate = NULL,
  cedate = NULL
)
}
\arguments{
\item{bdate}{Beginning date to check.
Year, month, day format.}

\item{edate}{Ending date to check.
Year, month, day format.}

\item{param}{Pollutant parameter that site is measuring.}

\item{minlat}{Minimum latitude coordinate.}

\item{maxlat}{Maximum latitude coordinate.}

\item{minlong}{Minimum longitude coordinate.}

\item{maxlong}{Maximum longitude coordinate.}

\item{cbdate}{Beginning date of last change to DB. (Optional)}

\item{cedate}{Ending date of last change to DB. (Optional)}
}
\value{
API response containing daily summary data bounded by lat long coords.
}
\description{
Returns daily summary data given a bounding box (lat, long).
}
\examples{
\dontrun{
bdate <- "20200101"
edate <- "20200102"
param <- "42401"
minlat <- 33.3
maxlat <- 33.6
minlong <- -87
maxlong <- -86.7
result <- get_daily_summary_in_bbox(bdate,
                                    edate,
                                    param,
                                    minlat,
                                    maxlat,
                                    minlong,
                                    maxlong)
result$Data
}
}
