% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_summary_service.R
\name{get_daily_summary_in_site}
\alias{get_daily_summary_in_site}
\title{Returns data summarized by day at measurement site level.}
\usage{
get_daily_summary_in_site(
  bdate,
  edate,
  state.fips,
  county,
  site,
  param,
  cbdate = NULL,
  cedate = NULL
)
}
\arguments{
\item{bdate}{Beginning date to check.
Year, month, day format.}

\item{edate}{Ending date to check.
Year, month, day format.}

\item{state.fips}{State FIPS code.
Use get_state_fips() if unsure.}

\item{county}{County code. 
Use get_counties_in_state() if unsure.}

\item{site}{Measurement site code.
Use get_sites_by_county() if unsure.}

\item{param}{Pollutant parameter that site is measuring.}

\item{cbdate}{Change begin date. (Optional)}

\item{cedate}{Change end date. (Optional)}
}
\value{
API response containing daily data.
}
\description{
Returns data summarized by day at measurement site level.
}
\examples{
\dontrun{
param <- 44201
bdate <- 20170618
edate <- 20170618
state <- 37
county <- 183
site <- 0014
result <- get_daily_summary_in_site(bdate,
                                    edate,
                                    state,
                                    county,
                                    site,
                                    param)
result$Data
}
}
