% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_summary_service.R
\name{get_daily_summary_in_state}
\alias{get_daily_summary_in_state}
\title{Returns daily data at the state level.}
\usage{
get_daily_summary_in_state(
  bdate,
  edate,
  state.fips,
  param,
  cached = TRUE,
  cache_directory = "/cache",
  cbdate = NULL,
  cedate = NULL
)
}
\arguments{
\item{bdate}{Beginning date to check.
Year, month, day format.}

\item{edate}{Ending date to check.
Year, month, day format.}

\item{state.fips}{State FIPS code.
Use get_state_fips() if unsure.}

\item{param}{Pollutant parameter that site is measuring.}

\item{cached}{TRUE or FALSE specifying if the data from the call is to 
be cached. Default: TRUE. (Optional)}

\item{cache_directory}{Place inside user-level cache directory to store 
the cached data. Default: "/cache". (Optional)}

\item{cbdate}{Change begin date. (Optional)}

\item{cedate}{Change end date. (Optional)}
}
\value{
API response containing daily data.
}
\description{
Returns daily data at the state level.
}
\examples{
\dontrun{
param <- 44201
bdate <- 20170618
edate <- 20170618
state <- 37
result <- get_daily_summary_in_state(bdate,
                                     edate,
                                     state,
                                     param)
result$Data
}
}
