% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_service.R
\name{get_fields_by_service}
\alias{get_fields_by_service}
\title{Get fields required per service.}
\usage{
get_fields_by_service(service)
}
\arguments{
\item{service}{A service provided by EPA's AQS system.}
}
\description{
Get fields required per service.
}
