% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_service.R
\name{get_known_issues}
\alias{get_known_issues}
\title{Get any known issues within the API.}
\usage{
get_known_issues()
}
\description{
Get any known issues within the API.
}
\examples{
\dontrun{
result <- get_known_issues()
result$Data
}
}
