% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitors_service.R
\name{get_monitors_in_site}
\alias{get_monitors_in_site}
\title{Get all monitors at a site.}
\usage{
get_monitors_in_site(
  bdate,
  edate,
  state.fips,
  county,
  param,
  site,
  cached = TRUE,
  cache_directory = "/cache"
)
}
\arguments{
\item{bdate}{Beginning date to check.
Year, month, day format.}

\item{edate}{Ending date to check.
Year, month, day format.}

\item{state.fips}{State FIPS code.
Use get_state_fips() if unsure.}

\item{county}{County code. 
Use get_counties_in_state() if unsure.}

\item{param}{Pollutant parameter that site is measuring.}

\item{site}{Measurement site code.
Use get_sites_by_county() if unsure.}

\item{cached}{TRUE or FALSE specifying if the data from the call is to 
be cached. Default: TRUE. (Optional)}

\item{cache_directory}{Place inside user-level cache directory to store 
the cached data. Default: "/cache". (Optional)}
}
\value{
API response containing operational information
about the monitor.
}
\description{
Get all monitors at a site.
}
\examples{
\dontrun{
bdate <- "20200101"
edate <- "20200102"
state.fips <- "37"
county <- "001"
param <- "42401"
site <- "001"
result <- get_monitors_in_site(bdate, edate, state.fips, county, param, site)
result$Data
} 
}
