% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_service.R
\name{get_parameter_classes}
\alias{get_parameter_classes}
\title{Get all types of parameters.}
\usage{
get_parameter_classes()
}
\value{
API response containing types of 
parameters and their respective code.
}
\description{
Get all types of parameters.
}
\examples{
\dontrun{
classes <- get_parameter_classes()
classes$Data
}
}
