% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_service.R
\name{get_parameters_in_class}
\alias{get_parameters_in_class}
\title{Get all parameters available within a particular parameter class.}
\usage{
get_parameters_in_class(class)
}
\arguments{
\item{class}{A type of pollutant. 
Find types of pollutants with get_parameter_classes().}
}
\value{
API response containing parameters 
found within a class/group of like parameters.
}
\description{
Get all parameters available within a particular parameter class.
}
\examples{
\dontrun{
class <- "AQI POLLUTANTS"
parameters <- get_parameters_in_class(class)
parameters$Data
}
}
