% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa_blanks_service.R
\name{get_qa_blanks_in_pqao}
\alias{get_qa_blanks_in_pqao}
\title{Get quality assurance blank data for a primary quality assurance
organization.}
\usage{
get_qa_blanks_in_pqao(bdate, edate, param, pqao)
}
\arguments{
\item{bdate}{Beginning date to check.
Year, month, day format.}

\item{edate}{Ending date to check.
Year, month, day format.}

\item{param}{Pollutant parameter that site is measuring.}

\item{pqao}{An encoding for a primary quality assurance organization.
If unsure, use get_all_pqaos().}
}
\value{
API response containing operational information
about the quality assurance blank data.
}
\description{
Get quality assurance blank data for a primary quality assurance
organization.
}
\examples{
\dontrun{
bdate <- "20180101"
edate <- "20180131"
param <- "88101"
pqao <- "0013"
result <- get_qa_blanks_in_pqao(bdate, edate, param, pqao)
result$Data
}
}
