% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa_blanks_service.R
\name{get_qa_blanks_in_state}
\alias{get_qa_blanks_in_state}
\title{Get quality assurance blank data in a state.}
\usage{
get_qa_blanks_in_state(bdate, edate, state.fips, param)
}
\arguments{
\item{bdate}{Beginning date to check.
Year, month, day format.}

\item{edate}{Ending date to check.
Year, month, day format.}

\item{state.fips}{State FIPS code.
Use get_state_fips() if unsure.}

\item{param}{Pollutant parameter that site is measuring.}
}
\value{
API response containing operational information
about the quality assurance blank data.
}
\description{
Get quality assurance blank data in a state.
}
\examples{
\dontrun{
bdate <- "20180101"
edate <- "20180131"
state.fips <- "01"
param <- "88101"
result <- get_qa_blanks_in_state(bdate, edate, state.fips, param)
result$Data
}
}
