% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa_ca_service.R
\name{get_qa_ca_in_site}
\alias{get_qa_ca_in_site}
\title{Get quality assurance collocated assessment data at
a site.}
\usage{
get_qa_ca_in_site(bdate, edate, state.fips, county, param, site)
}
\arguments{
\item{bdate}{Beginning date to check.
Year, month, day format.}

\item{edate}{Ending date to check.
Year, month, day format.}

\item{state.fips}{State FIPS code.
Use get_state_fips() if unsure.}

\item{county}{County code. 
Use get_counties_in_state() if unsure.}

\item{param}{Pollutant parameter that site is measuring.}

\item{site}{Measurement site code.
Use get_sites_by_county() if unsure.}
}
\value{
API response containing operational information
about the quality assurance annual performance evaluations.
}
\description{
Get quality assurance collocated assessment data at
a site.
}
\examples{
\dontrun{
bdate <- "20190101"
edate <- "20190131"
state.fips <- "01"
county <- "089"
param <- "88101"
site <- "0014"
result <- get_qa_ca_in_site(bdate, edate, state.fips, county, param, site)
result$Data
} 
}
