% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa_fra_service.R
\name{get_qa_fra_in_county}
\alias{get_qa_fra_in_county}
\title{Get quality assurance flow rate audit data in a county.}
\usage{
get_qa_fra_in_county(
  bdate,
  edate,
  state.fips,
  county,
  param,
  cached = TRUE,
  cache_directory = "/cache"
)
}
\arguments{
\item{bdate}{Beginning date to check.
Year, month, day format.}

\item{edate}{Ending date to check.
Year, month, day format.}

\item{state.fips}{State FIPS code.
Use get_state_fips() if unsure.}

\item{county}{County code. 
Use get_counties_in_state() if unsure.}

\item{param}{Pollutant parameter that site is measuring.}

\item{cached}{TRUE or FALSE specifying if the data from the call is to 
be cached. Default: TRUE. (Optional)}

\item{cache_directory}{Place inside user-level cache directory to store 
the cached data. Default: "/cache". (Optional)}
}
\value{
API response containing operational information
about the quality assurance flow rate audit data.
}
\description{
Get quality assurance flow rate audit data in a county.
}
\examples{
\dontrun{
bdate <- "20200101"
edate <- "20200131"
state.fips <- "01"
county <- "003"
param <- "88101"
result <- get_qa_fra_in_county(bdate, edate, state.fips, county, param)
result$Data
}
}
