% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa_fra_service.R
\name{get_qa_fra_in_pqao}
\alias{get_qa_fra_in_pqao}
\title{Get quality assurance flow rate audit data for a primary quality assurance
organization.}
\usage{
get_qa_fra_in_pqao(
  bdate,
  edate,
  param,
  pqao,
  cached = TRUE,
  cache_directory = "/cache"
)
}
\arguments{
\item{bdate}{Beginning date to check.
Year, month, day format.}

\item{edate}{Ending date to check.
Year, month, day format.}

\item{param}{Pollutant parameter that site is measuring.}

\item{pqao}{An encoding for a primary quality assurance organization.
If unsure, use get_all_pqaos().}

\item{cached}{TRUE or FALSE specifying if the data from the call is to 
be cached. Default: TRUE. (Optional)}

\item{cache_directory}{Place inside user-level cache directory to store 
the cached data. Default: "/cache". (Optional)}
}
\value{
API response containing operational information
about the quality assurance flow rate audit data.
}
\description{
Get quality assurance flow rate audit data for a primary quality assurance
organization.
}
\examples{
\dontrun{
bdate <- "20200101"
edate <- "20200131"
param <- "88101"
pqao <- "0013"
result <- get_qa_fra_in_pqao(bdate, edate, param, pqao)
result$Data
}
}
