% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarterly_service.R
\name{get_quarterly_summary_in_cbsa}
\alias{get_quarterly_summary_in_cbsa}
\title{Get quarterly summary data in a Core Based Statistical Area.}
\usage{
get_quarterly_summary_in_cbsa(
  bdate,
  edate,
  param,
  cbsa,
  cbdate = NULL,
  cedate = NULL
)
}
\arguments{
\item{bdate}{Beginning date to check.
Year, month, day format.}

\item{edate}{Ending date to check.
Year, month, day format.}

\item{param}{Pollutant parameter that site is measuring.}

\item{cbsa}{An encoding for a Core Base Statiscal Area.
If unsure, use get_cbsas().}

\item{cbdate}{Beginning date of last change to DB. (Optional)}

\item{cedate}{Ending date of last change to DB. (Optional)}
}
\value{
API response containing quarterly summary data at the cbsa level.
}
\description{
Get quarterly summary data in a Core Based Statistical Area.
}
\examples{
\dontrun{
bdate <- "20190101"
edate <- "20190601"
cbsa <- "16740"
param <- "42401"
result <- get_quarterly_summary_in_cbsa(bdate, 
                                        edate, 
                                        param, 
                                        cbsa)
result$Data
}
}
