% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarterly_service.R
\name{get_quarterly_summary_in_site}
\alias{get_quarterly_summary_in_site}
\title{Returns data summaries by yearly quarter.}
\usage{
get_quarterly_summary_in_site(
  bdate,
  edate,
  state.fips,
  county,
  param,
  site,
  cbdate = NULL,
  cedate = NULL
)
}
\arguments{
\item{bdate}{Beginning date to check.
Year, month, day format.}

\item{edate}{Ending date to check.
Year, month, day format.}

\item{state.fips}{State FIPS code.
Use get_state_fips() if unsure.}

\item{county}{County code. 
Use get_counties_in_state() if unsure.}

\item{param}{Pollutant parameter that site is measuring.}

\item{site}{Measurement site code.
Use get_sites_by_county() if unsure.}

\item{cbdate}{Beginning date of last change to DB. (Optional)}

\item{cedate}{Ending date of last change to DB. (Optional)}
}
\value{
API response containing quarterly summary data at a site.
}
\description{
Returns data summaries by yearly quarter.
}
\examples{
\dontrun{
bdate <- "20200101"
edate <- "20200102"
state.fips <- "37"
county <- "001"
site <- "001"
param <- "42401"
result <- get_quarterly_summary_in_site(bdate, 
                                        edate, 
                                        state.fips, 
                                        county, 
                                        param, 
                                        site)
result$Data
}
}
