% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_service.R
\name{get_revision_history}
\alias{get_revision_history}
\title{Get the API's revision history.}
\usage{
get_revision_history()
}
\description{
Get the API's revision history.
}
\examples{
\dontrun{
result <- get_revision_history()
result$Data
}
}
