% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_service.R
\name{get_samples_in_bbox}
\alias{get_samples_in_bbox}
\title{Get samples (finest grained data) for a bounding box (lat, long).}
\usage{
get_samples_in_bbox(
  bdate,
  edate,
  minlat,
  maxlat,
  minlong,
  maxlong,
  param,
  cached = TRUE,
  cache_directory = "/cache",
  duration = NULL,
  cbdate = NULL,
  cedate = NULL
)
}
\arguments{
\item{bdate}{Beginning date to check.
Year, month, day format.}

\item{edate}{Ending date to check.
Year, month, day format.}

\item{minlat}{Minimum latitude coordinate.}

\item{maxlat}{Maximum latitude coordinate.}

\item{minlong}{Minimum longitude coordinate.}

\item{maxlong}{Maximum longitude coordinate.}

\item{param}{Pollutant parameter that site is measuring.}

\item{cached}{TRUE or FALSE specifying if the data from the call is to 
be cached. Default: TRUE. (Optional)}

\item{cache_directory}{Place inside user-level cache directory to store 
the cached data. Default: "/cache". (Optional)}

\item{duration}{The 1-character AQS sample duration code. (Optional)}

\item{cbdate}{Beginning date of last change to DB. (Optional)}

\item{cedate}{Ending date of last change to DB. (Optional)}
}
\value{
API response containing operational information
about the monitor.
}
\description{
Get samples (finest grained data) for a bounding box (lat, long).
}
\examples{
\dontrun{
bdate <- "20200101"
edate <- "20200102"
param <- "42401"
minlat <- 33.3
maxlat <- 33.6
minlong <- -87
maxlong <- -86.7
result <- get_samples_in_bbox(bdate = bdate,
                              edate = edate,
                              param = param,
                              minlat = minlat,
                              maxlat = maxlat,
                              minlong = minlong,
                              maxlong = maxlong)
result$Data
}
}
