% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_service.R
\name{get_samples_in_cbsa}
\alias{get_samples_in_cbsa}
\title{Get samples (finest grained data) for a Core Based Statistical Area.}
\usage{
get_samples_in_cbsa(
  bdate,
  edate,
  param,
  cbsa,
  cached = TRUE,
  cache_directory = "/cache",
  duration = NULL,
  cbdate = NULL,
  cedate = NULL
)
}
\arguments{
\item{bdate}{Beginning date to check.
Year, month, day format.}

\item{edate}{Ending date to check.
Year, month, day format.}

\item{param}{Pollutant parameter that site is measuring.}

\item{cbsa}{An encoding for a Core Base Statiscal Area.
If unsure, use get_cbsas().}

\item{cached}{TRUE or FALSE specifying if the data from the call is to 
be cached. Default: TRUE. (Optional)}

\item{cache_directory}{Place inside user-level cache directory to store 
the cached data. Default: "/cache". (Optional)}

\item{duration}{The 1-character AQS sample duration code. (Optional)}

\item{cbdate}{Change begin date. (Optional)}

\item{cedate}{Change end date. (Optional)}
}
\value{
API response containing sample measurements in a CBSA.
}
\description{
Get samples (finest grained data) for a Core Based Statistical Area.
}
\examples{
\dontrun{
bdate <- "20200101"
edate <- "20200102"
cbsa <- "16740"
param <- "42401"
result <- get_samples_in_cbsa(bdate, edate, cbsa, param)
result$Data
}
}
