% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_service.R
\name{get_samples_in_county}
\alias{get_samples_in_county}
\title{Get samples (finest grained data) for a county.}
\usage{
get_samples_in_county(
  bdate,
  edate,
  state.fips,
  param,
  county,
  cached = TRUE,
  cache_directory = "/cache",
  duration = NULL,
  cbdate = NULL,
  cedate = NULL
)
}
\arguments{
\item{bdate}{Beginning date to check.
Year, month, day format.}

\item{edate}{Ending date to check.
Year, month, day format.}

\item{state.fips}{State FIPS code.
Use get_state_fips() if unsure.}

\item{param}{Pollutant parameter that site is measuring.}

\item{county}{County code.}

\item{cached}{TRUE or FALSE specifying if the data from the call is to 
be cached. Default: TRUE. (Optional)}

\item{cache_directory}{Place inside user-level cache directory to store 
the cached data. Default: "/cache". (Optional)}

\item{duration}{The 1-character AQS sample duration code. (Optional)}

\item{cbdate}{Change begin date. (Optional)}

\item{cedate}{Change end date. (Optional)}
}
\value{
API response containing sample measurements in a state.
}
\description{
Get samples (finest grained data) for a county.
}
\examples{
\dontrun{
bdate <- "20160101"
edate <- "20160102"
state.fips <- "15"
county <- "001"
param <- "42401"
result <- get_samples_in_county(bdate = bdate,
                                edate = edate,
                                param = param,
                                state.fips = state.fips,
                                county = county)
result$Data
} 
}
