% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_service.R
\name{get_samples_in_site}
\alias{get_samples_in_site}
\title{Get samples (finest grained data) for a measurement site.}
\usage{
get_samples_in_site(
  bdate,
  edate,
  state.fips,
  county,
  site,
  param,
  cached = TRUE,
  cache_directory = "/cache",
  duration = NULL,
  cbdate = NULL,
  cedate = NULL
)
}
\arguments{
\item{bdate}{Beginning date to check.
Year, month, day format.}

\item{edate}{Ending date to check.
Year, month, day format.}

\item{state.fips}{State FIPS code.
Use get_state_fips() if unsure.}

\item{county}{County code. 
Use get_counties_in_state() if unsure.}

\item{site}{Measurement site code.
Use get_sites_by_county() if unsure.}

\item{param}{Pollutant parameter that site is measuring.}

\item{cached}{TRUE or FALSE specifying if the data from the call is to 
be cached. Default: TRUE. (Optional)}

\item{cache_directory}{Place inside user-level cache directory to store 
the cached data. Default: "/cache". (Optional)}

\item{duration}{The 1-character AQS sample duration code. (Optional)}

\item{cbdate}{Change begin date. (Optional)}

\item{cedate}{Change end date. (Optional)}
}
\value{
API response containing samples at given site.
}
\description{
Get samples (finest grained data) for a measurement site.
}
\examples{
\dontrun{
bdate <- "20160101"
edate <- "20160102"
state.fips <- "15"
county <- "001"
param <- "42401"
site <- "0007"
cbdate <- "20200101"
cedate <- "20201231"
result <- get_samples_in_site(bdate = bdate, 
                              edate = edate, 
                              param = param, 
                              state.fips = state.fips,
                              county = county,
                              site = site)
result$Data
}
}
