% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_service.R
\name{get_samples_in_state}
\alias{get_samples_in_state}
\title{Get samples (finest grained data) for a state.}
\usage{
get_samples_in_state(
  bdate = bdate,
  edate = edate,
  state.fips = state.fips,
  param = param,
  cached = TRUE,
  cache_directory = "/cache",
  duration = NULL,
  cbdate = NULL,
  cedate = NULL
)
}
\arguments{
\item{bdate}{Beginning date to check.
Year, month, day format.}

\item{edate}{Ending date to check.
Year, month, day format.}

\item{state.fips}{State FIPS code.
Use get_state_fips() if unsure.}

\item{param}{Pollutant parameter that site is measuring.}

\item{cached}{TRUE or FALSE specifying if the data from the call is to 
be cached. Default: TRUE. (Optional)}

\item{cache_directory}{Place inside user-level cache directory to store 
the cached data. Default: "/cache". (Optional)}

\item{duration}{The 1-character AQS sample duration code. (Optional)}

\item{cbdate}{Change begin date. (Optional)}

\item{cedate}{Change end date. (Optional)}
}
\value{
API response containing sample measurements in a state.
}
\description{
Get samples (finest grained data) for a state.
}
\examples{
\dontrun{
bdate <- "20160101"
edate <- "20160102"
state.fips <- "15"
param <- "42401"
result <- get_samples_in_state(bdate = bdate, 
                               edate = edate, 
                               param = param, 
                               state.fips = state.fips)
result$Data
}
}
