% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_service.R
\name{get_sites_by_county}
\alias{get_sites_by_county}
\title{Get all measurement sites within a county.}
\usage{
get_sites_by_county(state.fips, county.code)
}
\arguments{
\item{state.fips}{A state FIPS code. 
Use get_state_fips() to find the appropriate FIPS code.}

\item{county.code}{A county code. 
Use get_counties_in_state() to find the appropriate code for a given county.}
}
\value{
API response containing all measurement sites within a given county.
}
\description{
Get all measurement sites within a county.
}
\examples{
\dontrun{
state <- "37"
county.code <- "001"
measurement.sites <- get_sites_by_county(state, county.code)
measurement.sites$Data 
}
}
