% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_service.R
\name{get_state_fips}
\alias{get_state_fips}
\title{Get all FIPS codes for each US state.}
\usage{
get_state_fips()
}
\value{
API response containing states and their respective FIPS codes.
}
\description{
Get all FIPS codes for each US state.
}
\examples{
\dontrun{
state.fips <- get_state_fips()
state.fips$Data
}
}
