% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_qa_ape_service.R
\name{get_tf_qa_ape_in_pqao}
\alias{get_tf_qa_ape_in_pqao}
\title{Get quality assurance annual performance 
evaluations for a primary quality assurance
organization in transaction format.}
\usage{
get_tf_qa_ape_in_pqao(bdate, edate, param, pqao)
}
\arguments{
\item{bdate}{Beginning date to check.
Year, month, day format.}

\item{edate}{Ending date to check.
Year, month, day format.}

\item{param}{Pollutant parameter that site is measuring.}

\item{pqao}{An encoding for a primary quality 
assurance organization.
If unsure, use get_all_pqaos().}
}
\value{
API response containing operational information
about the quality assurance annual performance 
evaluations in transaction format.
}
\description{
Get quality assurance annual performance 
evaluations for a primary quality assurance
organization in transaction format.
}
\examples{
\dontrun{
bdate <- "20170101"
edate <- "20171231"
pqao <- "0013"
param <- "44201"
result <- get_tf_qa_ape_in_pqao(bdate, edate, param, pqao)
result$Data
}
}
