% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_sample_service.R
\name{get_tf_sample_in_agency}
\alias{get_tf_sample_in_agency}
\title{Get sample data in the transaction format for a monitoring agency.}
\usage{
get_tf_sample_in_agency(
  bdate,
  edate,
  param,
  agency,
  cached = TRUE,
  cache_directory = "/cache"
)
}
\arguments{
\item{bdate}{Beginning date to check.
Year, month, day format.}

\item{edate}{Ending date to check.
Year, month, day format.}

\item{param}{Pollutant parameter that site is measuring.}

\item{agency}{The monitoring agency.}

\item{cached}{TRUE or FALSE specifying if the data from the call is to 
be cached. Default: TRUE. (Optional)}

\item{cache_directory}{Place inside user-level cache directory to store 
the cached data. Default: "/cache". (Optional)}
}
\value{
API response containing all sample data in submission format.
}
\description{
Get sample data in the transaction format for a monitoring agency.
}
\examples{
\dontrun{
bdate <- "20200101"
edate <- "20200331"
state.fips <- "01"
param <- "44201"
agency <- "0013"
result <- get_tf_sample_in_agency(bdate, 
                              edate, 
                              param, 
                              agency)
result$Data
}
}
