% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_sample_service.R
\name{get_tf_sample_in_site}
\alias{get_tf_sample_in_site}
\title{Get sample data in the transaction format at
a site.}
\usage{
get_tf_sample_in_site(
  bdate,
  edate,
  state.fips,
  county,
  param,
  site,
  cached = TRUE,
  cache_directory = "/cache"
)
}
\arguments{
\item{bdate}{Beginning date to check.
Year, month, day format.}

\item{edate}{Ending date to check.
Year, month, day format.}

\item{state.fips}{State FIPS code.
Use get_state_fips() if unsure.}

\item{county}{County code. 
Use get_counties_in_state() if unsure.}

\item{param}{Pollutant parameter that site is measuring.}

\item{site}{Measurement site code.
Use get_sites_by_county() if unsure.}

\item{cached}{TRUE or FALSE specifying if the data from the call is to 
be cached. Default: TRUE. (Optional)}

\item{cache_directory}{Place inside user-level cache directory to store 
the cached data. Default: "/cache". (Optional)}
}
\value{
API response containing all sample data in submission format.
}
\description{
Get sample data in the transaction format at
a site.
}
\examples{
\dontrun{
bdate <- "20200101"
edate <- "20200331"
state.fips <- "01"
county <- "003"
param <- "44201"
site <- "0010"
result <- get_tf_sample_in_site(bdate,
                                edate,
                                state.fips,
                                county,
                                param,
                                site)
result$Data
} 
}
