% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modified_rvest_function.R
\name{html_table.xml_node}
\alias{html_table.xml_node}
\title{Modified html_table.xml_node}
\usage{
\method{html_table}{xml_node}(x, header = NA, trim = TRUE, fill = FALSE,
  dec = ".")
}
\description{
See original specs for html_table.xml_node in rvest package. Modification done 
to allow for more flexible table scraping in websites.
}
