% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_placement.R
\name{is.API.running}
\alias{is.API.running}
\title{Check if the API is up and running}
\usage{
is.API.running()
}
\description{
Check if the API is up and running
}
\examples{
\dontrun{
is.API.running()
}
}
