% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_placement.R
\name{is_API_running}
\alias{is_API_running}
\title{Check if the API is up and running}
\usage{
is_API_running()
}
\description{
Check if the API is up and running
}
\examples{
\dontrun{
is_API_running()
}
}
