% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_placement.R
\name{list.cached.data}
\alias{list.cached.data}
\title{Shows contents of cache directory}
\usage{
list.cached.data(directory = "/cache")
}
\arguments{
\item{directory}{Place inside user-level cache directory that was used to store the cached data previously. Default: "/cache".}
}
\value{
Character vector of file names currently in cache directory.
}
\description{
Shows contents of cache directory
}
\examples{
\dontrun{
my.files <- list.cached.data()
my.files 
}
}
