% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transforms.R
\name{list.string.replacer}
\alias{list.string.replacer}
\title{Replace every string entry in a list}
\usage{
list.string.replacer(entry.list, pattern, replacement)
}
\arguments{
\item{entry.list}{List containing character entries}

\item{pattern}{Pattern to replace}

\item{replacement}{Replacement for entries following the pattern}
}
\value{
A list with entries matching the pattern replaced by replacement
}
\description{
Replace every string entry in a list
}
\examples{
\dontrun{
services <- epair:::get.services()
services <- epair:::list.string.replacer(services, "\\t", "")
services
}
}
