% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_utility.R
\name{lookup_by_bbox}
\alias{lookup_by_bbox}
\title{Internal function to perform geospatial lookup by bounding box (lat, long).}
\usage{
lookup_by_bbox(
  endpoint,
  bdate,
  edate,
  param,
  minlat,
  maxlat,
  minlong,
  maxlong,
  cached = TRUE,
  cache_directory = "/cache",
  duration = NULL,
  cbdate = NULL,
  cedate = NULL
)
}
\arguments{
\item{endpoint}{Base url to make call.}

\item{bdate}{Beginning date to check.
Year, month, day format.}

\item{edate}{Ending date to check.
Year, month, day format.}

\item{param}{Pollutant parameter that site is measuring.}

\item{minlat}{Minimum latitude coordinate.}

\item{maxlat}{Maximum latitude coordinate.}

\item{minlong}{Minimum longitude coordinate.}

\item{maxlong}{Maximum longitude coordinate.}

\item{cached}{TRUE or FALSE specifying if the data from the call is to 
be cached. Default: TRUE. (Optional)}

\item{cache_directory}{Place inside user-level cache directory to store 
the cached data. Default: "/cache". (Optional)}

\item{duration}{The 1-character AQS sample duration code. (Optional)}

\item{cbdate}{Beginning date of last change to DB. (Optional)}

\item{cedate}{Ending date of last change to DB. (Optional)}
}
\value{
API response containing bbox bounded data.
}
\description{
Internal function to perform geospatial lookup by bounding box (lat, long).
}
\examples{
\dontrun{
bdate <- "20200101"
edate <- "20200102"
param <- "42401"
minlat <- 33.3
maxlat <- 33.6
minlong <- -87
maxlong <- -86.7
result <- lookup_by_bbox(MONITORS, bdate, 
                                   edate, 
                                   param, 
                                   minlat, 
                                   maxlat, 
                                   minlong, 
                                   maxlong)
result$Data
}
}
