% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_utility.R
\name{lookup_by_cbsa}
\alias{lookup_by_cbsa}
\title{Internal function to perform geospatial lookup by Core Based Statistical Area.}
\usage{
lookup_by_cbsa(endpoint, bdate, edate, param, cbsa)
}
\arguments{
\item{endpoint}{Base url to make call.}

\item{bdate}{Beginning date to check.
Year, month, day format.}

\item{edate}{Ending date to check.
Year, month, day format.}

\item{param}{Pollutant parameter that site is measuring.}

\item{cbsa}{An encoding for a Core Base Statiscal Area.
If unsure, use get_cbsas().}
}
\value{
API response containing operational information
about the monitor.
}
\description{
Internal function to perform geospatial lookup by Core Based Statistical Area.
}
\examples{
\dontrun{
bdate <- "20200101"
edate <- "20200102"
cbsa <- "16740"
param <- "42401"
result <- lookup_by_cbsa(MONITORS, bdate, edate, param, cbsa)
result$Data
}
}
