% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_utility.R
\name{lookup_by_ma}
\alias{lookup_by_ma}
\title{Internal function to perform geospatial 
lookup by monitoring agency.}
\usage{
lookup_by_ma(endpoint, bdate, edate, param, agency)
}
\arguments{
\item{endpoint}{Base url to make call.}

\item{bdate}{Beginning date to check.
Year, month, day format.}

\item{edate}{Ending date to check.
Year, month, day format.}

\item{param}{Pollutant parameter that site is measuring.}

\item{agency}{The monitoring agency.}
}
\description{
Internal function to perform geospatial 
lookup by monitoring agency.
}
\examples{
\dontrun{
bdate <- "20200101"
edate <- "20201231"
param <- "44201"
agency <- "0013"
result <- lookup_by_ma(QA_APE, bdate, edate, param, agency)
result$Data
}
}
