% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_utility.R
\name{lookup_by_pqao}
\alias{lookup_by_pqao}
\title{Internal function to perform geospatial 
lookup by primary quality assurance organization.}
\usage{
lookup_by_pqao(endpoint, bdate, edate, param, pqao)
}
\arguments{
\item{endpoint}{Base url to make call.}

\item{bdate}{Beginning date to check.
Year, month, day format.}

\item{edate}{Ending date to check.
Year, month, day format.}

\item{param}{Pollutant parameter that site is measuring.}

\item{pqao}{An encoding for a Primary Quality Assurance Organization.
If unsure, use get_all_pqaos().}
}
\description{
Internal function to perform geospatial 
lookup by primary quality assurance organization.
}
\examples{
\dontrun{
bdate <- "20200101"
edate <- "20201231"
param <- "44201"
pqao <- "0013"
result <- lookup_by_pqao(QA_APE, bdate, edate, param, pqao)
result$Data
}
}
