% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_utility.R
\name{lookup_by_site}
\alias{lookup_by_site}
\title{Internal function to perform geospatial lookup by site.}
\usage{
lookup_by_site(endpoint, bdate, edate, state.fips, county, param, site)
}
\arguments{
\item{endpoint}{Base url to make call.}

\item{bdate}{Beginning date to check.
Year, month, day format.}

\item{edate}{Ending date to check.
Year, month, day format.}

\item{state.fips}{State FIPS code.
Use get_state_fips() if unsure.}

\item{county}{County code. 
Use get_counties_in_state() if unsure.}

\item{param}{Pollutant parameter that site is measuring.}

\item{site}{Measurement site code.
Use get_sites_by_county() if unsure.}
}
\value{
API response containing operational information
about the monitor.
}
\description{
Internal function to perform geospatial lookup by site.
}
\examples{
\dontrun{
bdate <- "20200101"
edate <- "20200102"
state.fips <- "37"
county <- "001"
site <- "001"
param <- "42401"
result <- lookup_by_site(MONITORS, bdate, edate, state.fips, county, param, site)
result$Data
}
}
