% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_utility.R
\name{lookup_by_state}
\alias{lookup_by_state}
\title{Internal function to perform geospatial lookup by state.}
\usage{
lookup_by_state(endpoint, bdate, edate, state.fips, param)
}
\arguments{
\item{endpoint}{Base url to make call.}

\item{bdate}{Beginning date to check.
Year, month, day format.}

\item{edate}{Ending date to check.
Year, month, day format.}

\item{state.fips}{State FIPS code.
Use get_state_fips() if unsure.}

\item{param}{Pollutant parameter that site is measuring.}
}
\value{
API response containing operational information
about the monitor.
}
\description{
Internal function to perform geospatial lookup by state.
}
\examples{
\dontrun{
bdate <- "20200101"
edate <- "20200102"
state.fips <- "37"
param <- "42401"
result <- lookup_by_state(MONITORS, bdate, edate, state.fips, param)
result$Data
}
}
