% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epa_scrapers.R
\name{non.endpoint.checker}
\alias{non.endpoint.checker}
\title{Check if a string contains characters not seen in endpoints}
\usage{
non.endpoint.checker(string)
}
\arguments{
\item{string}{A character entry from entries in the data frame of API services}
}
\value{
A boolean reflecting presence of endpoint in string.
}
\description{
Check if a string contains characters not seen in endpoints
}
\examples{
epair:::non.endpoint.checker("list/states")
epair:::non.endpoint.checker("https://example here")
}
