% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{perform.call}
\alias{perform.call}
\title{Perform call and convert data into list}
\usage{
perform.call(endpoint, variables = list(),
  name = deparse(substitute(variables)))
}
\arguments{
\item{call}{URL following structure from EPA API}
}
\value{
A list containing requested data
}
\description{
Perform call and convert data into list
}
