% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_placement.R
\name{place.call.raw}
\alias{place.call.raw}
\title{Perform call and maintain JSON Lite structure}
\usage{
place.call.raw(url)
}
\arguments{
\item{url}{URL following structure from EPA API}
}
\value{
Results of data request in JSON format
}
\description{
Perform call and maintain JSON Lite structure
}
\examples{
\dontrun{
endpoint <- 'list/states'
call <- create.base.call(endpoint)
raw.call <- place.call.raw(call)
raw.call
}
}
