% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_placement.R
\name{place.call}
\alias{place.call}
\title{Place the URL as a call to the EPA API}
\usage{
place.call(url)
}
\arguments{
\item{url}{A string with a valid URL for the EPA API}
}
\value{
Result of query from the API
}
\description{
Place the URL as a call to the EPA API
}
\examples{
\dontrun{
url <- "user_url"
result <- place.call(url)
}
}
