% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/making_services.R
\name{populate.all.services}
\alias{populate.all.services}
\title{Turn tables of API services into a list}
\usage{
populate.all.services(tables.to.modify)
}
\arguments{
\item{tables.to.modify}{List of tables from API. Each table is a data frame.}
}
\value{
A list with each service and filters as chained variables to make for easy calling.
}
\description{
Turn tables of API services into a list
}
\examples{
\dontrun{
tables.to.modify <- get.all.tables()
services <- populate.all.services(tables.to.modify)
services$List
}
}
