% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transforms.R
\name{remove.escapes.spaces}
\alias{remove.escapes.spaces}
\title{Remove tabs, new lines, and empty spaces from entries in a data frame}
\usage{
remove.escapes.spaces(df)
}
\arguments{
\item{df}{Data frame to remove tabs, new lines, and empty spaces from}
}
\value{
Data frame without tabs, new lines, and empty spaces
}
\description{
Remove tabs, new lines, and empty spaces from entries in a data frame
}
\examples{
service <- c("Sign up")
description <- c("Email will\r\n\t\t\t\t\t\t\t be sent to the registered
address from aqsdatamart@epa.gov.")

og.df <- data.frame(service, description)
clean.df <- remove.escapes.spaces(og_table)
clean.df
}
