% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_placement.R
\name{retrieve.cached.call}
\alias{retrieve.cached.call}
\title{Retrieves memory of previously cached call.}
\usage{
retrieve.cached.call(endpoint, variables = list(), directory = "/cache")
}
\arguments{
\item{endpoint}{An endpoint from the available EPA API endpoints}

\item{variables}{A list of variables or a single variable to filter the EPA API endpoint.}

\item{directory}{Place inside user-level cache directory that was used to store the cached data previously. Default: "/cache".}
}
\value{
A list containing requested data
}
\description{
Retrieves memory of previously cached call.
}
\examples{
\dontrun{
endpoint <- 'list/states'
retrieve.cached.call(endpoint)
}
}
