% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_placement.R
\name{save.new.cached.call}
\alias{save.new.cached.call}
\title{Saves a new call that has not been previously cached yet.}
\usage{
save.new.cached.call(endpoint, variables = list(), user.path)
}
\arguments{
\item{endpoint}{An endpoint from the available EPA API endpoints}

\item{variables}{A list of variables or a single variable to filter the EPA API endpoint.}

\item{directory}{Place inside user-level cache directory that was used to store the cached data previously. Default: "~/epair/cache".}
}
\value{
A list containing requested data
}
\description{
Saves a new call that has not been previously cached yet.
}
\examples{
\dontrun{
endpoint <- 'list/states'
save.new.cached.call(endpoint)
}
}
