% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/making_services.R
\name{setup.service}
\alias{setup.service}
\title{Make list of single service}
\usage{
setup.service(df)
}
\arguments{
\item{df}{Data frame with info to make an API service.}
}
\value{
A list with the filter content of a service set to the service name.
}
\description{
Make list of single service
}
\examples{
\dontrun{
tbls <- get.all.tables()
single <- tbls[[8]]
setup.service(single)
}
}
