% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/making_services.R
\name{setup.single.filter}
\alias{setup.single.filter}
\title{Create a single filter}
\usage{
setup.single.filter(filter.name, i, df)
}
\arguments{
\item{filter.name}{Name of filter in API service}

\item{i}{Row number to use to create filter. Make sure filter information is 
present at i before hand.}

\item{df}{Data frame with filter information.}
}
\value{
A list with filter content given to the filter name.
}
\description{
Create a single filter
}
\examples{
\dontrun{
tbls <- get.all.tables()
single <- tbls[[9]]
filter.name <- "My filter"
setup.single.filter(filter.name, 1, single)
}
}
