% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transforms.R
\name{string.replacer}
\alias{string.replacer}
\title{Replace all characters entries in data frame}
\usage{
string.replacer(df, pattern, replacement)
}
\arguments{
\item{df}{Data frame containing character entries}

\item{pattern}{Pattern to use for matching}

\item{replacement}{Replacement of entries matching pattern}
}
\value{
A data frame with entries following the pattern being replaced by replacement
}
\description{
Replace all characters entries in data frame
}
\examples{
df <- data.frame(c("1", "2", "3", "4"))
modified.df <- epair:::string.replacer(df, "1", "One")
modified.df
}
