% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets.R
\docType{data}
\name{variable.types}
\alias{variable.types}
\title{Variable parameter names to use}
\description{
The variable.types list contains the listing endpoints for finding out more
information in making calls requiring more variables.
}
\keyword{datasets}
