% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets.R
\docType{data}
\name{variables}
\alias{variables}
\title{Variables used for querying in EPA API}
\description{
The variables data frame contains information about what variables can be used
to build queries in the EPA API.
}
\keyword{datasets}
