% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/making_services.R
\name{change.classes.filter}
\alias{change.classes.filter}
\title{Update name to parameter class entry}
\usage{
change.classes.filter(services)
}
\arguments{
\item{services}{List of services offered by the API.}
}
\value{
Services with corrected name of filter for parameter classes.
}
\description{
As of this package release, a service in the API was incorrectly described
in the original website. This function gives the services variable the proper information
describing usage of the service.
}
\examples{
\dontrun{
services <- get.services()
services <- change.classes.filter(services)
services$List$Filter$`Parameter Classes (groups of parameters, like criteria or all)`
}
}
