% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transforms.R
\name{get.transpose}
\alias{get.transpose}
\title{Transpose a data frame}
\usage{
get.transpose(df)
}
\arguments{
\item{df}{Data frame to be transposed}
}
\value{
The transposed data frame. First variable entries become column names.
}
\description{
Transpose a data frame
}
\examples{
\dontrun{
url <- "https://aqs.epa.gov/aqsweb/documents/data_api.html"
table.path <- '//*[@id="main-content"]/div[2]/div[1]/div/div/table[1]'
df <- epair:::get.table(url, table.path)
t.df <- epair:::get.transpose(df)
t.df
}
}
