% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitors_service.R
\name{get_monitors_in_county}
\alias{get_monitors_in_county}
\title{Get all monitors in a county.}
\usage{
get_monitors_in_county(bdate, edate, state.fips, county, param)
}
\arguments{
\item{bdate}{Beginning date to check.
Year, month, day format.}

\item{edate}{Ending date to check.
Year, month, day format.}

\item{state.fips}{State FIPS code.
Use get_state_fips() if unsure.}

\item{county}{County code. 
Use get_counties_in_state() if unsure.}

\item{param}{Pollutant parameter that site is measuring.}
}
\value{
API response containing operational information
about the monitor.
}
\description{
Get all monitors in a county.
}
\examples{
\dontrun{
bdate <- "20200101"
edate <- "20200102"
state.fips <- "37"
county <- "001"
param <- "42401"
result <- get_monitors_in_county(bdate, edate, state.fips, county, param)
result$Data
}
}
