% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitors_service.R
\name{get_monitors_in_state}
\alias{get_monitors_in_state}
\title{Get monitors in state.}
\usage{
get_monitors_in_state(bdate, edate, state.fips, param)
}
\arguments{
\item{bdate}{Beginning date to check.
Year, month, day format.}

\item{edate}{Ending date to check.
Year, month, day format.}

\item{state.fips}{State FIPS code.
Use get_state_fips() if unsure.}

\item{param}{Pollutant parameter that site is measuring.}
}
\value{
API response containing operational information
about the monitor.
}
\description{
Get monitors in state.
}
\examples{
\dontrun{
bdate <- "20200101"
edate <- "20200102"
state.fips <- "37"
param <- "42401"
result <- get_monitors_in_state(bdate, edate, state.fips, param)
result$Data
}
}
