% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa_qc_service.R
\name{get_qa_qc_in_pqao}
\alias{get_qa_qc_in_pqao}
\title{Get quality assurance one point quality control data
for a primary quality assurance
organization.}
\usage{
get_qa_qc_in_pqao(bdate, edate, param, pqao)
}
\arguments{
\item{bdate}{Beginning date to check.
Year, month, day format.}

\item{edate}{Ending date to check.
Year, month, day format.}

\item{param}{Pollutant parameter that site is measuring.}

\item{pqao}{An encoding for a primary quality assurance organization.
If unsure, use get_all_pqaos().}
}
\value{
API response containing operational information
about the quality assurance one point quality control data.
}
\description{
Get quality assurance one point quality control data
for a primary quality assurance
organization.
}
\examples{
\dontrun{
bdate <- "20170101"
edate <- "20171231"
pqao <- "0013"
param <- "44201"
result <- get_qa_qc_in_pqao(bdate, edate, param, pqao)
result$Data
}
}
