% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{perform.call.raw}
\alias{perform.call.raw}
\title{Perform call and keep original result}
\usage{
perform.call.raw(endpoint, variables = list(),
  name = deparse(substitute(variables)))
}
\arguments{
\item{call}{URL following structure from EPA API}
}
\value{
A list containing result from query to EPA API
}
\description{
Perform call and keep original result
}
