% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_placement.R
\name{perform.call}
\alias{perform.call}
\title{Cached version of the perform.call function}
\usage{
perform.call(
  endpoint,
  variables = list(),
  directory = "~/epair/cache",
  cached = TRUE
)
}
\arguments{
\item{endpoint}{An endpoint from the available EPA API endpoints}

\item{variables}{A list of variables or a single variable to filter the EPA API endpoint.}

\item{cached}{TRUE or FALSE specifying if the data from the call is to be cached. Default: TRUE.}

\item{cache_directory}{Place inside user-level cache directory to store the cached data. Default: "~/epair/cache".}
}
\value{
A list containing requested data
}
\description{
Cached version of the perform.call function
}
\examples{
\dontrun{
endpoint <- 'list/states'
result <- perform.call(endpoint)
}
}
