% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_placement.R
\name{perform.call}
\alias{perform.call}
\title{Perform call and convert data into list}
\usage{
perform.call(endpoint, variables = list(),
  name = deparse(substitute(variables)))
}
\arguments{
\item{endpoint}{An endpoint from the available EPA API endpoints}

\item{variables}{A list of variables or a single variable to filter the EPA API endpoint.}

\item{name}{Specifies the name each variable should have when placed in the URL.}
}
\value{
A list containing requested data
}
\description{
Perform call and convert data into list
}
\examples{
\dontrun{
endpoint <- 'list/states'
result <- perform.call(endpoint)
}
}
