% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{remove.escapes.spaces}
\alias{remove.escapes.spaces}
\title{Remove '\t', '\r\n', "   " from entries in a data frame}
\usage{
remove.escapes.spaces(df)
}
\arguments{
\item{df}{Data frame to remove '\t', '\r\n', "   " from}
}
\value{
Data frame without '\t', '\r\n', "   "
}
\description{
Remove '\t', '\r\n', "   " from entries in a data frame
}
\examples{
url <- "https://aqs.epa.gov/aqsweb/documents/data_api.html"
table.path <- '//*[@id="main-content"]/div[2]/div[1]/div/div/table[1]'
df <- get.table( url, table.path )
df

clean.df <- remove.escapes.spaces( df )
clean.df
}
