% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transforms.R
\name{remove.escapes.spaces}
\alias{remove.escapes.spaces}
\title{Remove tabs, new lines, and empty spaces from entries in a data frame}
\usage{
remove.escapes.spaces(df)
}
\arguments{
\item{df}{Data frame to remove tabs, new lines, and empty spaces from}
}
\value{
Data frame without tabs, new lines, and empty spaces
}
\description{
Remove tabs, new lines, and empty spaces from entries in a data frame
}
\examples{
\dontrun{
url <- "https://aqs.epa.gov/aqsweb/documents/data_api.html"
table.path <- '//*[@id="main-content"]/div[2]/div[1]/div/div/table[1]'
df <- epair:::get.table(url, table.path)
df

clean.df <- epair:::remove.escapes.spaces(df)
clean.df
}
}
